Deduplicate validation/suggestion code
