+++ dataverse-webapp/src/main/java/edu/harvard/iq/dataverse/authorization/providers/builtin/DataverseUserPage.java
@@ -8,12 +8,11 @@ import edu.harvard.iq.dataverse.authorization.AuthenticationProvider;
 import edu.harvard.iq.dataverse.authorization.AuthenticationServiceBean;
 import edu.harvard.iq.dataverse.authorization.EditableAccountField;
 import edu.harvard.iq.dataverse.authorization.UserRecordIdentifier;
+import edu.harvard.iq.dataverse.authorization.providers.common.BaseUserPage;
 import edu.harvard.iq.dataverse.authorization.providers.shib.ShibAuthenticationProvider;
 import edu.harvard.iq.dataverse.common.BundleUtil;
 import edu.harvard.iq.dataverse.consent.ConsentDto;
 import edu.harvard.iq.dataverse.consent.ConsentService;
-import edu.harvard.iq.dataverse.dataset.metadata.inputRenderer.Suggestion;
-import edu.harvard.iq.dataverse.dataset.metadata.inputRenderer.suggestion.SuggestionHandler;
 import edu.harvard.iq.dataverse.mail.confirmemail.ConfirmEmailException;
 import edu.harvard.iq.dataverse.mail.confirmemail.ConfirmEmailServiceBean;
 import edu.harvard.iq.dataverse.mail.confirmemail.ConfirmEmailUtil;
@@ -34,10 +33,7 @@ import edu.harvard.iq.dataverse.settings.SettingsServiceBean;
 import edu.harvard.iq.dataverse.settings.SettingsWrapper;
 import edu.harvard.iq.dataverse.util.JsfHelper;
 import edu.harvard.iq.dataverse.util.SystemConfig;
-import edu.harvard.iq.dataverse.validation.OrcIdValidator;
 import edu.harvard.iq.dataverse.validation.PasswordValidatorServiceBean;
-import edu.harvard.iq.dataverse.validation.RorValidator;
-import edu.harvard.iq.dataverse.validation.field.ValidationResult;
 import io.vavr.control.Option;
 import org.apache.commons.lang.StringUtils;
 import org.hibernate.validator.constraints.NotBlank;
@@ -56,7 +52,6 @@ import java.io.UnsupportedEncodingException;
 import java.net.URLDecoder;
 import java.sql.Timestamp;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.Date;
 import java.util.List;
 import java.util.Locale;
@@ -66,7 +61,7 @@ import java.util.logging.Logger;
 
 @ViewScoped
 @Named("DataverseUserPage")
-public class DataverseUserPage implements java.io.Serializable {
+public class DataverseUserPage extends BaseUserPage {
 
     private static final Logger logger = Logger.getLogger(DataverseUserPage.class.getCanonicalName());
 
@@ -96,12 +91,6 @@ public class DataverseUserPage implements java.io.Serializable {
     SystemConfig systemConfig;
     @EJB
     PasswordValidatorServiceBean passwordValidatorService;
-    @EJB
-    RorValidator rorValidator;
-    @EJB
-    OrcIdValidator orcIdValidator;
-    @EJB(beanName = "RorSuggestionHandler")
-    SuggestionHandler suggestionHandler;
     @Inject
     SettingsWrapper settingsWrapper;
     @Inject
@@ -347,37 +336,6 @@ public class DataverseUserPage implements java.io.Serializable {
         if (!errors.isEmpty()) {
             ((UIInput) toValidate).setValid(false);
         }
-    }
-
-    public void validateOrcId(FacesContext context, UIComponent toValidate, Object value) {
-        String orcid = (String) value;
-        if (StringUtils.isEmpty(orcid)) {
-            return;
-        }
-        ValidationResult result = orcIdValidator.validate(orcid);
-        if (!result.isOk()) {
-            ((UIInput) toValidate).setValid(false);
-            context.addMessage(toValidate.getClientId(context), new FacesMessage(FacesMessage.SEVERITY_ERROR,
-                    BundleUtil.getStringFromBundle("user." + result.getErrorCode()), null));
-        }
-    }
-
-    public void validateAffiliationRor(FacesContext context, UIComponent toValidate, Object value) {
-        String ror = (String) value;
-        if (StringUtils.isEmpty(ror)) {
-            return;
-        }
-
-        ValidationResult result = rorValidator.validate(ror);
-        if (!result.isOk()) {
-            ((UIInput) toValidate).setValid(false);
-            context.addMessage(toValidate.getClientId(context), new FacesMessage(FacesMessage.SEVERITY_ERROR,
-                    BundleUtil.getStringFromBundle("user.affiliationror." + result.getErrorCode()), null));
-        }
-    }
-
-    public List<Suggestion> processAffiliationRorSuggestions(String query) {
-        return suggestionHandler.generateSuggestions(Collections.emptyMap(), query);
     }
 
     public String save() {
