+++ dataverse-webapp/src/main/java/edu/harvard/iq/dataverse/authorization/providers/common/ExternalIdpFirstLoginPage.java
@@ -16,8 +16,6 @@ import edu.harvard.iq.dataverse.authorization.providers.saml.SamlAuthenticationS
 import edu.harvard.iq.dataverse.common.BundleUtil;
 import edu.harvard.iq.dataverse.consent.ConsentDto;
 import edu.harvard.iq.dataverse.consent.ConsentService;
-import edu.harvard.iq.dataverse.dataset.metadata.inputRenderer.Suggestion;
-import edu.harvard.iq.dataverse.dataset.metadata.inputRenderer.suggestion.SuggestionHandler;
 import edu.harvard.iq.dataverse.notification.NotificationObjectType;
 import edu.harvard.iq.dataverse.notification.UserNotificationService;
 import edu.harvard.iq.dataverse.persistence.config.EMailValidator;
@@ -30,10 +28,7 @@ import edu.harvard.iq.dataverse.settings.InstallationConfigService;
 import edu.harvard.iq.dataverse.settings.SettingsWrapper;
 import edu.harvard.iq.dataverse.util.JsfHelper;
 import edu.harvard.iq.dataverse.util.SystemConfig;
-import edu.harvard.iq.dataverse.validation.OrcIdValidator;
 import edu.harvard.iq.dataverse.validation.PasswordValidatorServiceBean;
-import edu.harvard.iq.dataverse.validation.RorValidator;
-import edu.harvard.iq.dataverse.validation.field.ValidationResult;
 import io.vavr.control.Option;
 import org.apache.commons.lang3.StringUtils;
 import org.hibernate.validator.constraints.NotBlank;
@@ -48,7 +43,6 @@ import javax.inject.Inject;
 import javax.inject.Named;
 import javax.servlet.http.HttpSession;
 import java.io.IOException;
-import java.io.Serializable;
 import java.sql.Timestamp;
 import java.util.ArrayList;
 import java.util.Collections;
@@ -70,7 +64,7 @@ import java.util.stream.Collectors;
  */
 @Named("ExternalIdpFirstLoginPage")
 @SessionScoped
-public class ExternalIdpFirstLoginPage implements Serializable {
+public class ExternalIdpFirstLoginPage extends BaseUserPage {
 
     private static final Logger logger = Logger.getLogger(ExternalIdpFirstLoginPage.class.getCanonicalName());
 
@@ -94,15 +88,6 @@ public class ExternalIdpFirstLoginPage implements Serializable {
 
     @EJB
     InstallationConfigService installationConfigService;
-
-    @EJB
-    RorValidator rorValidator;
-
-    @EJB
-    OrcIdValidator orcIdValidator;
-
-    @EJB(beanName = "RorSuggestionHandler")
-    SuggestionHandler suggestionHandler;
 
     @Inject
     DataverseSession session;
@@ -328,37 +313,6 @@ public class ExternalIdpFirstLoginPage implements Serializable {
                     BundleUtil.getStringFromBundle("user.email.taken"), null);
             context.addMessage(toValidate.getClientId(context), message);
         }
-    }
-
-    public void validateOrcId(FacesContext context, UIComponent toValidate, Object value) {
-        String orcid = (String) value;
-        if (org.apache.commons.lang.StringUtils.isEmpty(orcid)) {
-            return;
-        }
-        ValidationResult result = orcIdValidator.validate(orcid);
-        if (!result.isOk()) {
-            ((UIInput) toValidate).setValid(false);
-            context.addMessage(toValidate.getClientId(context), new FacesMessage(FacesMessage.SEVERITY_ERROR,
-                    BundleUtil.getStringFromBundle("user." + result.getErrorCode()), null));
-        }
-    }
-
-    public void validateAffiliationRor(FacesContext context, UIComponent toValidate, Object value) {
-        String ror = (String) value;
-        if (org.apache.commons.lang.StringUtils.isEmpty(ror)) {
-            return;
-        }
-
-        ValidationResult result = rorValidator.validate(ror);
-        if (!result.isOk()) {
-            ((UIInput) toValidate).setValid(false);
-            context.addMessage(toValidate.getClientId(context), new FacesMessage(FacesMessage.SEVERITY_ERROR,
-                    BundleUtil.getStringFromBundle("user.affiliationror." + result.getErrorCode()), null));
-        }
-    }
-
-    public List<Suggestion> processAffiliationRorSuggestions(String query) {
-        return suggestionHandler.generateSuggestions(Collections.emptyMap(), query);
     }
 
     public String getWelcomeMessage() {
